@@ -5,7 +5,6 @@ import {
   signOut,
 } from "firebase/auth";
 import { useEffect, useState } from "react";
-import { useAlert } from "react-alert";
 import Initialization from "./Initialization";
 
 Initialization();
@@ -14,7 +13,6 @@ const useFirebase = () => {
   const auth = getAuth();
   const [user, setUser] = useState(null);
   const [loading, setLoading] = useState(true);
-  const alart = useAlert();
 
   const LogInWithEmain = (email, password) => {
     return signInWithEmailAndPassword(auth, email, password);
@@ -23,7 +21,7 @@ const useFirebase = () => {
   const logOut = () => {
     signOut(auth)
       .then(() => {
-        alart.show("You singed out");
+        alert("You singed out");
       })
       .catch((err) => console.log(err.message));
   };
